Fix entity temp ID being 32 bits on the server (but only 16 on the client).

+++ src/common/net/Deserializer.h
@@ -39,7 +39,7 @@ struct BaseEventData {
  */
 struct BaseEntityData {
   quint32 id;
-  quint32 temp_id;
+  quint16 temp_id;
   QPoint position;
 };
